Fix congestion and build errors

+++ ATSPM/Data/Models/SpeedManagement/CongestionTracking/CongestionSeriesData.cs
@@ -4,6 +4,6 @@ namespace ATSPM.Data.Models.SpeedManagement.CongestionTracking
     public class CongestionSeriesData
     {
         public List<DataPoint<double>> Average {  get; set; }
-        public List<DataPoint<int>> EightyFifth {  get; set; }
+        public List<DataPoint<long>> EightyFifth {  get; set; }
     }
 }
